Ownership verification failed Verification method: HTML tag Failure reason: Your meta tag is not in the <head> section of your home page. Please fix your implementation and reverify, or use another verification method.
Copy and paste the AdSense code in between the <head></head> tags of your site
Place this code on every page across your site and Google will automatically show ads in all the best places for you

pls fix and add
<script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-6805990598879954"
     crossorigin="anonymous"></script>
It can take up to an hour for ads to appear on the page. See our code implementation guide⁠‎ for more details.